5.7. Ипотека [#156315]

+++ src/test/java/ru/job4j/loop/MortgageTest.java
@@ -7,7 +7,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 class MortgageTest {
     @Test
     void whenLoan100Salary120Percent20ThenYear1() {
-        int loan = 100;
+        double loan = 100;
         int salary = 120;
         double percent = 0.20;
         int expected = 1;
@@ -17,7 +17,7 @@ class MortgageTest {
 
     @Test
     void whenLoan100Salary50Percent20ThenYear3() {
-        int loan = 100;
+        double loan = 100;
         int salary = 50;
         double percent = 0.20;
         int expected = 3;
@@ -27,7 +27,7 @@ class MortgageTest {
 
     @Test
     void whenLoan150Salary50Percent20ThenYear6() {
-        int loan = 150;
+        double loan = 150;
         int salary = 50;
         double percent = 0.20;
         int expected = 6;
